refactor(msw): separate mock data and handlers into distinct folders
