+++ packages/msw/src/index.ts
@@ -1,14 +1,2 @@
-export {
-  handlers,
-  authHandlers,
-  walletHandlers,
-  keycloakHandlers,
-  mockUser,
-  mockWallets,
-  mockTokens,
-  mockTransfers,
-  mockKeycloakToken,
-} from "./handlers";
-
-// server is exported from its sub-path:
-// @treetracker/msw/server  → Node.js / Jest
+export * as handlers from "./handlers";
+export * as mocks from "./mocks";
