+++ packages/msw/src/mocks/auth.ts
@@ -1,0 +1,5 @@
+export const mockUser = {
+  userId: "mock-user-id",
+  username: "mockuser",
+  email: "mock@example.com",
+};
