+++ packages/msw/src/mocks/keycloak.ts
@@ -1,0 +1,6 @@
+export const mockKeycloakToken = {
+  access_token: "mock-keycloak-access-token",
+  expires_in: 3600,
+  token_type: "Bearer",
+  scope: "openid",
+};
